数据库字段类型修改

+++ src/main/java/xyz/playedu/api/mapper/UserLearnDurationStatsMapper.java
@@ -13,7 +13,7 @@ import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 @Mapper
 public interface UserLearnDurationStatsMapper extends BaseMapper<UserLearnDurationStats> {
 
-    Integer getUserDuration(Integer userId);
+    Long getUserDuration(Integer userId);
 
 }
